Build: Remove unused `BMQVersionUtil.cmake` module
This module computes the version of the broker being built by searching the
source code file `mqbscm_versiontag.h` for three version macros, and parsing
out the values they are set to.  There are two issues with this:

  1. The normal CMake way to get the version is to set the version in the
     `project()` macro, which we do not do.  This would not be so bad if we
     manually called the functions in this module to set version variables,
     but…

  2. This module is completely unused.

The correct way to handle this situation (I believe) is to generate
`mqbscm_versiontag.h` using CMake’s `configure_file` command, and to store the
source-of-truth version information in `project()`.

This patch removes the unused CMake module.  A future patch will perform this
generation.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>