Build: Generate MQB version from CMake information
Continuing on from the previous commit, this patch removes the existing
`mqbscm_versiontag.h` file and instead generates it from the CMake project
version information.  For the moment, this has no effect.

Ideally, we would want this generated file to live in the build directory,
rather than in the source tree.  There are, however, a few issues with this
approach.  First, because we depend on the bde-tools build system, we don’t
easily have the way to modify the way we generate the private package targets
for mqbscm.  There may be a way to hack around it, but for simplicity we
haven’t gone that route here.  Second, doing so would involve mucking around
with the include directories for mqb and all dependent targets; this again
needs to be done carefully.  Third and finally, this keeps the Doxygen
documentation building without further configuration.  In the future, perhaps
we will be able to move this file and the `*_cpp03.{cpp,h}` files (and
eventually maybe the XSD-generated files) into the build directory, rather than
have them placed in the source directory.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>