Fix: Remove all uses of `bmqsys::ThreadUtil::k_SUPPORT_THREAD_NAME`
The contracts of `bmqsys::ThreadUtil::setCurrentThreadName` and
`bmqsys::ThreadUtil::setCurrentThreadNameOnce` are wide; they may be called on
any platform, and they have no side effects on platforms where setting a thread
name is unsupported.  This means we do not need to guard calls to them by
checking a constant that is only true on platforms where setting a thread name
is supported.  This patch removes all such guards, which removes all uses of
this constant.

In most cases, the use of this constant is in an `if` statement that obviously
wraps some call to `bmqsys::ThreadUtil::setCurrentThreadName` or
`bmqsys::ThreadUtil::setCurrentThreadNameOnce`.  The only less obvious case is
in `m_bmqbrkr_task`’s mtrap support, where this constant guards a write of the
command `k_MTRAP_SET_THREADNAME` to the mtrap pipe.  The command is handled by
`Task::onControlMessage` in the same file, which in turn calls
`bmqsys::ThreadUtil::setCurrentThreadName`.  Presumably, this is where the `if`
guard should have been, but now it is safe to remove the guard on the write.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>