Feat: Add `integration-tests` target to CMake
This patch adds our Python pytest-based integration tests as first-class CTest
test cases, so users do not need to manually set up an environment and run the
`run-tests` script from any particular directory.  With our integration tests
being CTest test cases, it is now possible to run a single integration test
from an IDE that understands CTest test cases, such as VSCode with the
[vscode-cmake-tools][vscode-cmake-tools] extension, or Emacs with
[projectile][emacs-projectile] or [projection][emacs-projection].  After
applying this patch, users should be able to simply configure the project and
build the `integration-tests` target to build the broker and run all its
integration tests.

In order to do this, this patch first sets up a Python 3 virtual environment in
the CMake build directory that contains all the development requirements to use
our integration test framework.  Then, this patch uses that virtual environment
to add the integration tests as CTest targets, with the label `integration`.
This allows users to run only the integration tests, without running the unit
tests as well.

This patch introduces two optional dependencies during the CMake configuration.
If Python 3 is found, a virtual environment is created, and all requirements in
`requirements.txt` are installed into that virtual environment.  Second, an
additional dependency on the Python package `cmake-pytest` is added; if Python
3 is found and this package can be installed, then we are able to add the
`integration-tests` target.

[vscode-cmake-tools]: https://github.com/microsoft/vscode-cmake-tools
[emacs-projectile]: https://github.com/bbatsov/projectile
[emacs-projection]: https://github.com/mohkale/projection

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>