Build[bmqeval]: Turn warnings off in generated code
The code that is generated from Bison and Flex does not pass the aggressive
warnings that Clang outputs; this pollutes the build log substantially in a way
that we can’t really fix.  This patch turns off the warnings just these
components.  There are two parts of this solution: first, we need to turn off
the warnings in the build system for this small package; and second, we need to
use pragmas to turn off warnings only in the generated Bison header, which is
included in several other translation units, and thus which we cannot turn
warnings off for in the same way as above.

The warnings chosen here to turn off are the minimum needed to stop the
generated code from firing any warnings.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>