CI: Run build action on integration branches
Currently, the GitHub Action `build.yaml` is set to run on pushes and PRs that
target the `main` branch.  This severely restricts our ability to make changes
against longer-lived integration branches that may not be released immediately.
This patch relaxes that restriction, so we can run CI builds on pushes and PRs
that target branches beginning with the prefix `integration/`.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>