Fix: Remove unused exception objects
Clang emits two warnings about unused exception objects when building the
project now:

    src/applications/bmqstoragetool/m_bmqstoragetool_compositesequencenumber.cpp:77:45: warning: unused exception parameter 'e' [-Wunused-exception-parameter]
       77 |         catch (const bsl::invalid_argument& e) {
          |                                             ^

These exceptions are the prescribed way `bsl::stoul` signals errors, but
the exception objects are unused.  We already log a unique message based
on the result, so we have no need to name the exception object.  Rather
than mark it unused with an annotation, we simple remove its name.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>