Fix: Remove incorrect `BSLA_UNUSED` annotations
Clang warns us that there are four parameters that are marked as unused,
but are in fact used.  This is an example warning:

    src/applications/bmqstoragetool/m_bmqstoragetool_searchresult.cpp:557:48: warning: 'recordIndex' was marked unused but was used [-Wused-but-marked-unused]
      557 |                                                recordIndex,
          |                                                ^

In each of these cases, the parameter is forwarded on to other
functions which do not use the value.  This means that the parameter is
in fact used in each place this warning fires.  This patch fixes this
situation by removing the four incorrect `BSLA_UNUSED` annotations.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>