Fix: Mark unused allocators as unused
Clang warns us about a few allocator data members that are not used:

    In file included from src/applications/bmqstoragetool/m_bmqstoragetool_cslsearchresult.cpp:17:
    src/applications/bmqstoragetool/m_bmqstoragetool_cslsearchresult.h:98:23: warning: private field 'd_allocator_p' is not used [-Wunused-private-field]
       98 |     bslma::Allocator* d_allocator_p;
          |                       ^

We have two options to fix these.  For one, we could remove the
allocator data members, and have the allocator parameters to the
classes’ constructors be marked as `BSLA_UNUSED`.  This patch takes the
other option, by keeping these data members and marking them as
`BSLA_UNUSED`.  It is not clear to me that either is much better in each
of the cases fixed by this patch.  The solution chosen here at least is
a more minimal code change, will let us use the allocator some day, and
will alert us when we start to use the allocator.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>