Feat: Always handle SIGTERM and SIGQUIT gracefully
SIGTERM is the standard signal to politely tell a running program to
terminate.

Previously, SIGTERM (and SIQUIT) only resulted in gracefully shutdown
when running the broker in a TTY. This is not ideal behavior.

Now, the broker always performs a graceful shutdown upon receiving these
signals. This behavior now aligns with the intended use of these signals.

Signed-off-by: Christopher Beard <cbeard9@bloomberg.net>