Fix use sigkill to stop broker in fuzz test driver
Fuzz tests do not exit happily/pass if there are leaks during graceful
shutdown. We should fix that, but it's out of scope for this particular
change.

This commit uses SIGKILL to kill the broker after running fuzz tests
instead of SIGTERM to keep the same broker termination behavior (in the
context of the broker now handling SIGTERM gracefully).

Signed-off-by: Christopher Beard <cbeard9@bloomberg.net>