Fix: Fix warnings in CRC32 code
The CRC32 computation code generates a few warnings, which point to real
potential bugs in.  This patch cleans some of these up, and documents
others.

First, the code switches needlessly between signed and unsigned types.
This causes some signed-unsigned comparison warnings.  Second, the code
crucially relies on the word size of the platform being 64 bits or
fewer.  While this is true on all modern platforms, we add an assert
just in case.  Thirdly, this patch explicitly documents what happens
when a completely aligned buffer is passed in, and fixes a warning about
the `switch` statement not being exhaustive, which is intentional.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>