Fix: Explicitly cast signed-unsigned conversions
This patch adds two explicit casts to the `put` function of both
`FileBackedStorage` and `InMemoryStorage`, silencing warnings from
Clang, and making it explicit that we have a type boundary between one
interface that provides us with a signed `int`, and another that
provides us with an `unsigned int`.  A cursory examination of both of
these interfaces does not give a clear answer to whether we can (or
should) change their type; instead we make explicit the unsafe
type-conversion behavior here.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>