Fix: Avoid signed-unsigned warnings in tests
This patch fixes the last of the signed-unsigned comparison warnings in
our unit tests.  We mostly do this by changing the types that are being
compared, rather than introducing new casts.  However, when we need to
introduce a new cast, we choose to cast the reference value we are
comparing to, rather than the return value from the
component-under-test.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>