Refactor[bmqa]: Remove trivially true asserts from test
This patch removes two asserts from the `bmqa_message` unit test that
are now trivially true due to the previous warning fixes.  One is a
nonnegative check that is now trivially true by virtue of the type now
being unsigned; one is a postcondition check that is necessarily true
now.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>