Fix: allow graceful exit in fuzz test driver
Previously, fuzz tests expected the broker to exit with a non-zero
return code when terminated "as intended" because SIGTERM was not
handled.

Now, the broker gracefully exits with 0 upon succesful graceful shutdown
upon SIGTERM. This commit fixes the process management to handle this
new termination behavior gracefully.

Note: fuzz tests are currently running in CI with the
StackTraceTestAllocator, which aborts if a leak is found at termination.
In this case, the fuzz test will fail.

Signed-off-by: Christopher Beard <cbeard9@bloomberg.net>