Fix: Require latency monitoring domain to match prefix
Latency monitoring domains are an under-documented feature of
BlazingMQ that can be used to measure end-to-end message latency
through the broker.  To support partition-aware latency measurement,
the broker parses queue names on these domains using this
regex (src/groups/mqb/mqbc/mqbc_clusterstate.cpp:765):

    ^\S+\.([0-9]+)\.\S+\.\S+$

An example string that matches this is:

    foo.2.bar.baz

If a queue name matches this regex, the integer in the second
component (e.g., `2` in `foo.2.bar.baz`) is treated as the target
partition, and the queue is assigned to that partition.

Latency monitoring domains are expected to start with a configured
prefix (by default `bmq.sys.latemon.latency`, see
src/groups/mqb/mqbcfg/mqbcfg.xsd:102).  However, the existing
implementation matches the configured prefix *anywhere* within the
domain name.  This patch fixes this behavior: domains are not
recognized as latency monitoring domains only if they begin with the
configured prefix.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>