Refactor[MQBA]: Remove `Response` and `DomainConfigRaw` uses
The `mqbconfm::Response` type stores the result of looking up a
domain’s configuration on disk, either as a failure or as the raw
config, parsed from the JSON file on disk.  It is generated from a
type in `mqbconf.xsd`, which otherwise describes the actual domain
configuration file format.  Luckily, this type is only used
transiently in one component: it stores a cached result in
`mbqa::ConfigProvider`, and is passed between functions in
`mqba::ConfigProvider` as vocubulary.

Because the type is so simple, though, we prefer instead to just store
the domain configuration in cache and passing the domain configuration
directly between functions.  Doing so allows us to remove both
`mqbconfm::Response` and `mqbconfm::Failure` from `mqbconf.xsd`.

Furthermore, doing so shows that the type `mqbconfm::DomainConfigRaw`
is used in exactly the same way, in exactly the same places, and is
also a trivial wrapper type.  Eventually, through the storage and
wrapping via `mqbconfm::Response` and `mqbconfm::DomainConfigRaw`, we
evenutally read out the domain configuration as a `bsl::string` and
pass that to a user-specified callback—all within the one component
`mbqa_configprovider`.

This patch performs the refactoring within `mbqa_configprovider` to
allow us to remove `mqbconfm::Response` and
`mqbconfm::DomainConfigRaw`.  These types are replaced with the simple
`bsl::string`s that they represent.  This patch maintains the existing
behavior of the config provider, including:

  - Not adding failure results to the cache
  - Returning the same error codes
  - Reading from the domain configuration file with exactly the same
    code.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>