docs: Generate Doxygen tag files
Tag files are a feature of Doxygen that enable other projects to
generate documentation that links into our published generated
documentation.  For example, a project with a BlazingMQ consumer that
exposes `bmqa::Message` in the interface can be configured to link to
our published documentation for `bmqa::Message`.  It’s good practice
to generate these.

This patch turns on the generation of tagfiles in our documentation
targets.  By default, they are named `tags.xml` and put in the output
directory for each of our documentation targets.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>