fix: Fix Doxygen warnings from README
Since the inception of the project, we have used the repository’s README
as the main page for our generated Doxygen documentation.  Doxygen
doesn’t understand the HTML we use in our README, though, so it
generates warnings on every build.

This patch adds a very simple main page to our Doxygen documentation,
written in Doxygen’s native format (a file with the extension `.dox`
that contains a Doxygen-style C++ comment).  This new main page replaces
our README only in the generated API documentation, not on our GitHub
repo.  After applying this patch, there should be no warnings when
generating our API documentation.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>