Feat[mqb]: Add broker stats to stat log
This commit adds the # of clients and queues to the broker's stat log.
These values can be derived from existing stats with some legwork;
logging them in one line just makes it easier.

Note: These metrics show the number of connected clients and opened
queues. If a queue exists in storage but has no clients, it does not
count.

Signed-off-by: Christopher Beard <cbeard9@bloomberg.net>