Fix: Make narrowing conversion from `Int64` to `int` explicit
This patch fixes a warning from GCC about narrowing an `Int64` to an
`int` in `bmqimp::Application`:

    conversion from ‘BloombergLP::bsls::Types::Int64’ {aka ‘long long int’} to ‘int’ may change value [-Wconversion]

In `SessionOptions`, we take time durations as BDE `TimeDuration`s,
which represent time intervals with nanosecond precision, using a pair
of `Int64`s.  Our code ignores this the nanosecond precision and only
uses seconds.

In the particular case of the `statsDumpInterval()`, we copy the time
interval in seconds into an `int`, which results in a narrowing
conversion.  Rather than changing the internal `int`s, which are in the
stats layer and may need to be `int`s, we choose to explicitly cast from
`Int64` to `int` before the stats layer.  In this case, doing so is
safe, because `SessionOptions` enforces that `statsDumpInterval()` must
smaller than 60 minutes, which easily fits in an `int`.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>