Fix: Remove narrowing conversion from `Int64` to `int`
This patch fixes a warning from GCC about narrowing an `Int64` to an
`int` in `bmqimp::Application`:

    conversion from ‘BloombergLP::bsls::Types::Int64’ {aka ‘long long int’} to ‘int’ may change value [-Wconversion]

In `SessionOptions`, we take time durations as BDE `TimeDuration`s,
which represent time intervals with nanosecond precision, using a pair
of `Int64`s.  Our code ignores this the nanosecond precision and only
uses seconds.

In the particular case of keeping track of the next time to dump stats, we copy
the time interval in seconds from `statsDumpInterval()` into an `int` named
`d_nextStatDump`, which results in a narrowing conversion.  This `int` is
functionally just a counter, so there is no need for it to have a different
type than `statsDumpInterval()`.  Rather than introduce an explicit cast, we
change the type of `d_nextStatDump` to `Int64` to avoid the conversion.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>