Fix: Make max message data size explicit as precondition
`bmqa::Message` has two ways of setting a message’s payload:

  1. Passing a `Blob`, whose length is `int`, or
  2. Passing a pointer-to-data and a `size_t` length.

Both cases forward to `PutEventBuilder`’s `setMessagePayload`, which
takes an `int` and is a public interface, so it’s not easy to change.
Furthermore, under the hood, `PutEventBuilder` packs a raw pointer and
length into a `Blob` (`PutEventBuilder::packMessage`) whose length is an
`int`.  For the `Blob` overload, there is no issue, but for the pointer
and `size_t` overload, there is an implicit narrowing conversion from
`size_t` to the `int` that `Blob` uses as a length.  Ideally, this
interface would never have been a `size_t`, but it seems we’re stuck
with that decision.

This patch makes the requirement that the length passed in as a `size_t`
is representable as an `int` explicit as a precondition, and makes the
narrowing cast explicit in the code.  This fixes a warning reported by
GCC:

    conversion from ‘size_t’ {aka ‘long unsigned int’} to ‘int’ may change value [-Wconversion]

and also warns users of previously undocumented undefined behavior.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>