Fix[bmqimp]: Close connection when reopen times out
This patch forces the channel to close when an operation times out while
a queue is reopening.  In this state, we do not know whether the queue
has been successfully reopened or whether it will be in the future if
the remote broker catches up.  The safest thing to do here is to kill
the connection to the broker and try to reconnect anew.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>