UT[mqbblp]: Move virtual method definitions out-of-line
This patch fixes two newly-introduced warnings in the
`mqbblp_queueconsumptionmonitor` UT, which have the following shape:

    src/groups/mqb/mqbblp/mqbblp_queueconsumptionmonitor.t.cpp:113:8: warning: 'MockStorageIterator' has no out-of-line virtual method definitions; its vtable will be emitted in every translation unit [-Wweak-vtables]
      113 | struct MockStorageIterator : public mqbi::StorageIterator {
          |        ^

This warning is a bit of a false warning here, because these classes are
restricted to just the test file.  However, the fix of moving the definitions
of the methods for the affected classes out-of-line is easy, so we prefer that.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>