CI: Make apidocs push failures an error
Our API documentation CI workflow has been failing since September due
to an unintended change in the `gh-pages` branch protection settings.
We didn’t catch this, because the API documentation CI workflow is
configured to invariably pass, even if any command fails.  This change
was made in commit 844ba5c as part of a fix for adding untracked
generated documentation, but that fix was overzealous.

This patch makes any failure to push to `gh-pages` a CI error.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>