Feat[bmqtool]: Let bmqtool select session timeouts
In order to write integration tests for timeouts in a reasonable way, we
need the ability to set the timeouts in `bmqt::SessionOptions` on
sessions created with bmqtool.  This patch adds an optional CLI argument
to bmqtool `--timeout`, which takes as its single argument the number of
seconds to set all timeouts in `bmqt::SessionOptions` to.  This defaults
to the same default as `bmqt::SessionOptions`, namely 300 seconds.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>