Fix[bmqimp]: Close connection when open times out
This patch forces the channel to close when an operation times out while a
queue is opening.  In this state, we do not know whether the queue has been
successfully reopened or whether it will be in the future if the remote broker
catches up.  The safest thing to do here is to kill the connection to the
broker and let the user try to open the queue again.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>