Fix[bmqimp]: Close connection when close times out
This patch forces the channel to close when an operation times out while a
queue is being closed.  In this state, we do not know whether the queue has
been successfully closed or whether it will be in the future if the remote
broker catches up.  The safest thing to do here is to kill the connection to
the broker, closing this (and all queues) and letting queues that were open be
automatically reopened when we reconnect to a broker.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>