Fix[mqbblp]: Fix build on Solaris
This patch fixes a build error on Solaris/C++03: the new `ElectorEventFunctor`
constructor takes a `bslmf::MovableRef<bmqp::Event>`, which in C++11 is an
alias for `bmqp::Event&&`.  In C++03, though, temporary objects are not
automatically converted into `MovableRef`s, resulting in an invocation of the
`createInplace` function template to fail instantiation.  We solve this by
explicitly calling `bslmf::MovableRefUtil::move` on this temporary, which is a
no-op in C++11 but a necessary call in C++03.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>