Refactor[MQBA]: Remove `DomainResolver` generated type uses
The `mqbconfm::DomainResolver` type stores the result of resolving a
domain, following any redirects in the domain configuration.  It is
generated from a type in `mqbconf.xsd`, which otherwise describes the
actual domain configuration file format.  Luckily, this type is only
used transiently in one component: it stores a cached result in
`mbqa::DomainResolver`, and is passed between functions in
`mqba::DomainResolver` as vocubulary.  This is the same pattern as
`mqbconfm::Response` and `mqbconfm::DomainConfigRaw`, removed in an
earlier commit.

Because the type is so simple, though, we prefer instead to just store
the domain resolution result in cache and passing the domain
resolution result directly between functions.  Doing so allows us to
remove `mqbconfm::DomainResolver` from `mqbconf.xsd`.

This patch performs the refactoring within `mbqa_domainresolver` to
allow us to remove `mqbconfm::DomainResolver`.  This type is replaced
with two simple `bsl::string`s that it represents.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>