IT[test_client_timeout]: Ensure queues are assigned
The `test_client_timeout.py` integration test creates producers with
artificially short timeouts in order to test the behavior of clients
on open, reopen, and close timeouts.  However, because bmqtool only
supports a single timeout flag, the short timeouts can cause the
initial open queue operation to fail on slow test runners.  This patch
introduces a client with a long timeout to each test, which connects
and opens the necessary queues, assigning them on the primary.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>