Feat[bmqp]: Add `userAgent` to `ClientIdentity`
The BlazingMQ project provides three SDKs for writing applications
that send and receive messages from a BlazingMQ cluster:

  - the C++03 `libbmq` library, housed in this repository;
  - the Java `com.bloomberg.bmq` package; and
  - the Python `blazingmq` package.

Of these, `libbmq` and `com.bloomberg.bmq` are separate, independent
implementations of the BlazingMQ protocol, and `blazingmq` is a Cython
wrapper around `libbmq`.  Future SDKs, produced either by the
BlazingMQ project or by the community, may take either road.

It is useful telemetry to know which SDK language and version a client
is connecting to a broker with, and the `ClientIdentity` generated
type does partially provide this information, in the form of an
`sdkLanguage` enumeration and an `sdkVersion` integer.  These two
fields are useful for programmatic checks on protocol compatibility,
but for telemetry, we have found them to be insufficient.

First, any SDK that wraps `libbmq`, like the Python `blazingmq`
package, will advertise itself as an `E_CPP` client.  `libbmq`
provides no way for a client to override the SDK language and version
it negotiates with, so `blazingmq` has no way to advertise itself as
anything different.  While it is useful to know what version of the
underlying `libbmq` a Python client is connecting with, this leaves it
difficult for us to distinguish between clients that are connecting
directly using `libbmq`, or Python clients that are connecting using
`blazingmq`.  This problem will be true for any future SDK that wraps
`libbmq`.

Second, any new SDK that does not wrap `libbmq` will either need to
choose from one of the existing `sdkLanguages` (or use `E_UNKNOWN`),
or request a protocol change from BlazingMQ to add their language
version.  The same problem as above happens with the first option,
while the second option is simply unscalable.

This patch introduces a new optional field called `userAgent` to the
`ClientIdentity` type that is sent during session negotiation.  At the
protocol level, this field is a free-form string, with which client
libraries can identify themselves, and the broker does nothing more
than log it along with the session negotiation message.  This allows
clients that do not wrap `libbmq` to identify themselves to the
broker.

This patch also teaches `libbmq` and `bmqbrkr` about this field, so
they can send identifying version strings as their `userAgent`.  At
the moment, `libbmq` does provide any way to modify this field, so
libraries that wrap `libbmq`, like the Python `blazingmq`, still
suffer from the first problem noted above.  A future patch will expose
a way to modify the user agent (likely in restricted ways).

Importantly, this field MUST NOT be used for anything more than
telemetry, because clients do not need to specify it, and malicious
client libraries can easily spoof their `userAgent` to be the same as
other client libraries.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>