Feat: Add `s_versionDotString` to `*scm_version` components
This patch adds a new field `s_versionDotString` to the `Version`
structs in `*scm_version` components, which stores a human-readable
version string.  This differs from the `version()` function, which has
information identifying the package group or application prefixed to
it.  In principle, `s_versionDotString` could contain other
information, like prerelease numbers if we adopt those, without
breaking an compatibility for existing callers of `version()`.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>