Fix: MovableRef on C++03
Pull request #1037 introduced a build regression for C++03 compilers:

    blazingmq/src/groups/bmq/bmqc/bmqc_multiqueuethreadpool.h", line 999: Error: The operation "! BloombergLP::bslmf::MovableRef<bsl::shared_ptr<TestItem>>" is illegal.

This expression `! movableRef` is expanded from
`BSLS_ASSERT_SAFE(movableRef)`.  In C++11 and later,
`bslmf::MovableRef<T>` is a type alias for `T&&`, so an expression of
the form `! movableRef` is well-defined, when `T` supports `operator!`.
In C++03, though, `MovableRef<T>` is a type template which does not
support `operator!`.

This patch fixes this by calling `bslmf::MovableRefUtil::access`
function in the assertion expression.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>