ci: Remove date from dependency cache
We key our cache entries for dependencies with the hash of the
`build-deps.sh` script and the date the dependencies are built on.
Because we run this job nightly, we end up with a new cache entry each
night containing the same build results.  This quickly fills up our
cache with duplicate entries, crowding out space for the cache entries
we use to communicate build results between jobs.

In fact, it is very unusual to key a cache on the date in GitHub
Actions.  Most languages encourage you to use the hash of a dependency
lock file (à la `yarn.lock`) for repeatability.  In our case, the the
`build-deps.sh` contains the BDE and NTF tags we need to build.  As
long as we use stable version tags or specific commit SHAs, each
workflow run always tries to build the same source code, building the
same binaries.

This patch removes the date from the cache entry name.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>