chore: Use ntf-core with C++03 build support on Linux/Darwin
The version of ntf-core that build against in CI prevents Linux and
Darwin users from selecting what C++ standard to build with, instead
forcing the latest C++ standard that NTF supports.  Not only does this
block us from adding a CI job for building BlazingMQ with C++03 in
GitHub Actions, this is not even what we want with our existing C++17
builds.

[Pull request #353 to ntf-core][ntf353] fixes this behavior, so that
we can correctly select what C++ standard to build NTF with.  This PR
is included with the most recent versions of NTF, so this patch bumps
our minimum NTF version to the most recent version in every place we
build it.

However, the most recent versions of NTF have also added default
dependencies on three compression libraries in support of
functionality we do not use in BlazingMQ.  Of these three compression
libraries, BlazingMQ already depends on zlib, so this one added
dependency is harmless.  However, the other two compression libraries
zstd and lz4 are not dependencies of BlazingMQ, and we do not install
them in CI.  Just bumping the version of NTF will result in build
failures.  So, this patch also explicitly configures NTF without zstd
and lz4, so we do not introduce any additional dependencies to our
build.

In addition, the newest version of NTF supports TLS using OpenSSL.  As
we are preparing to add TLS support to BlazingMQ, this additional
dependency will be needed very soon, so this patch also installs
OpenSSL where it’s needed.

[ntf353]: https://github.com/bloomberg/ntf-core/pull/353

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>